Rename `check` package to `graph`
