+++ graph/search.go
@@ -1,4 +1,4 @@
-package check
+package graph
 
 import (
 	"fmt"
